# apps/docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Scriora Developer API",
    "description": "Universal B2B Omnichannel Social Publishing, Scheduling, and Human Governance Engine — Canonical REST Gateway.",
    "version": "1.0.0",
    "contact": {
      "name": "Scriora Developer Relations",
      "url": "https://scriora.io/docs",
      "email": "api@scriora.io"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User JWT session token obtained via /v1/auth/login or /v1/auth/register"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Workspace developer API key (sk_live_...)"
      }
    },
...
  },
  "paths": {
...
    "/v1/posts/{postId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "postId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/posts/{postId}/cancel": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "postId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development Server (Port 4000)"
    },
    {
      "url": "https://api.scriora.io",
      "description": "Production Global Gateway (Edge CDN)"
    }
  ],
...
}